refactor: organize employee_mock.go into a folder
